own_projects/wine/wine_recommendations_with_llm/requirements.txt
langchain==0.0.277
pinecone-client==2.2.2
streamlit==1.26.0
pandas
google-generativeai